- (Diep) - Moved arena exclusive files to an arena folder and renamed them appropriately.
