+++ src/app/diep/ui/main-menu/diep.arena-checkbox-renderer.ts
@@ -1,4 +1,4 @@
-import { TileType } from '../../engine/subsystems/diep.arena-manager';
+import { TileType } from '../../engine/subsystems/arena/arena.manager';
 import { DiepButton } from '../../core/diep.interfaces';
 import { DiepButtonAnimator } from '../buttons/diep.button-animator';
